refactor: Split comparison operators into separate modules
- Create comparison/ directory with modular structure
- Extract equality operators into equality.rs (=, <>)
- Extract ordering operators into ordering.rs (<, <=, >, >=)
- Keep shared comparison logic and type coercion in mod.rs
- Maintain public API compatibility with ComparisonOps struct
- All existing tests pass

Closes #1302
